+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Improvements:
 * [OLMIS-7319](https://openlmis.atlassian.net/browse/OLMIS-7319): Improved the login and forgot password screens to be responsive also at very small resolutions.
 * [OLMIS-7318](https://openlmis.atlassian.net/browse/OLMIS-7318): Added the ability to hide unnecessary modules on mobile.
 * [OLMIS-7314](https://openlmis.atlassian.net/browse/OLMIS-7314): Update scss files to enable webpack build.
+* [OLMIS-7318](https://openlmis.atlassian.net/browse/OLMIS-7318): Add possibility to display menu items only on mobile.
 
 
 6.2.6 / 2021-05-27